+++ Tests/YCoreUITests/Components/ElevationTests.swift
@@ -9,38 +9,6 @@
 import XCTest
 @testable import YCoreUI
 
-struct Elevation {
-    let offset: CGSize
-    let blur: CGFloat
-    let spread: CGFloat
-    let color: UIColor
-    let opacity: CGFloat
-    let useShadowPath: Bool
-    
-    init(
-        offset: CGSize,
-        blur: CGFloat,
-        spread: CGFloat,
-        color: UIColor,
-        opacity: CGFloat,
-        useShadowPath: Bool = true
-    ) {
-        self.offset = offset
-        self.blur = blur
-        self.spread = spread
-        self.color = color
-        self.opacity = opacity
-        self.useShadowPath = useShadowPath
-    }
-    
-    func apply(layer: CALayer, cornerRadius: CGFloat) {
-        guard useShadowPath else { return }
-        
-        let rect = layer.bounds.insetBy(dx: -spread, dy: -spread)
-        layer.shadowPath = UIBezierPath(roundedRect: rect, cornerRadius: cornerRadius).cgPath
-    }
-}
-
 final class ElevationTests: XCTestCase {
     func test_init_deliversGivenValues() {
         let sut = makeSUT()
